% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BacherTCellData.R
\name{BacherTCellData}
\alias{BacherTCellData}
\title{Obtain the Bacher T cell data}
\usage{
BacherTCellData(filtered = TRUE, ensembl = FALSE, location = TRUE)
}
\arguments{
\item{filtered}{Logical scalar indicating whether to filter out cells that were not used by the authors.}

\item{ensembl}{Logical scalar indicating whether the output row names should contain Ensembl identifiers.}

\item{location}{Logical scalar indicating whether genomic coordinates should be returned.}
}
\value{
A \linkS4class{SingleCellExperiment} object with a single matrix of UMI counts.
}
\description{
Obtain the human COVID T cell single-cell RNA-seq dataset from Bacher et al. (2020).
}
\details{
Column metadata is scraped from GEO, using both the author-supplied TSV of per-cell annotations and the sample-level metadata.
This contains information such as the diagnosis, severity, WHO class, clustering and clonotype.

If \code{filtered=TRUE}, only the cells used by the authors in their final analysis are returned.
Otherwise, an additional \code{filtered} field will be present in the \code{\link{colData}}, indicating whether the cell was retained by the authors. 

If \code{ensembl=TRUE}, the gene symbols are converted to Ensembl IDs in the row names of the output object.
Rows with missing Ensembl IDs are discarded, and only the first occurrence of duplicated IDs is retained.

If \code{location=TRUE}, the coordinates of the Ensembl gene models are stored in the \code{\link{rowRanges}} of the output.

All data are downloaded from ExperimentHub and cached for local re-use.
Specific resources can be retrieved by searching for \code{scRNAseq/bacher-tcell}.
}
\examples{
sce <- BacherTCellData()

}
\references{
Bacher P et al. (2020). 
Low avidity T cell responses to SARS-CoV-2 in unexposed individuals and severe COVID-19
\emph{Immunity} 53, 1258-1271
}
\author{
Aaron Lun
}
