% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_bib.R
\name{datasetsPage}
\alias{datasetsPage}
\title{use DT::datatable to present information in listDatasets() with linked HTML for pubmed entries}
\usage{
datasetsPage()
}
\description{
use DT::datatable to present information in listDatasets() with linked HTML for pubmed entries
}
