% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERCCSpikeInConcentrations.R
\name{ERCCSpikeInConcentrations}
\alias{ERCCSpikeInConcentrations}
\title{Obtain the ERCC concentration from Thermofisher}
\usage{
ERCCSpikeInConcentrations(genes = NULL, volume_nl = NULL, dilution = NULL)
}
\arguments{
\item{genes}{A character vector specifying the ERCC spike-in genes used.}

\item{volume_nl}{Volume (in nanoliters) of ERCC spike-in mixture.}

\item{dilution}{The ratio of ERCC mixture to total well content.}
}
\value{
A \code{data.frame} object with ERCC spike-in information.
}
\description{
Obtain the ERCC concentration from Thermofisher
}
\details{
All data are downloaded from ExperimentHub and cached for local re-use.
Specific resources can be retrieved by searching for \code{scRNAseq/ercc-spike-in-concentrations}.
}
\examples{
sce <- ERCCSpikeInConcentrations()

}
\author{
Alan O'Callaghan
}
