% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchDataset.R
\name{fetchDataset}
\alias{fetchDataset}
\title{Fetch a dataset from the gypsum backend}
\usage{
fetchDataset(name, version, subpath = NULL, package = "scRNAseq", ...)
}
\arguments{
\item{name}{String containing the name of the dataset.}

\item{version}{String containing the version of the dataset.}

\item{subpath}{String containing the path to a subdataset, if \code{name} contains multiple datasets.}

\item{package}{String containing the name of the package.}

\item{...}{Further arguments to pass to \code{\link{readObject}}.}
}
\value{
\code{fetchDataset} returns the dataset as a \linkS4class{SummarizedExperiment} or one of its subclasses.

\code{fetchMetadata} returns a named list of metadata for the specified dataset.
}
\description{
Fetch a dataset (or its metadata) from the gypsum backend.
}
\author{
Aaron Lun
}
