% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrunHSCData.R
\name{GrunHSCData}
\alias{GrunHSCData}
\title{Obtain the Grun HSC data}
\usage{
GrunHSCData(ensembl = FALSE, location = TRUE)
}
\arguments{
\item{ensembl}{Logical scalar indicating whether the output row names should contain Ensembl identifiers.}

\item{location}{Logical scalar indicating whether genomic coordinates should be returned.}
}
\value{
A \linkS4class{SingleCellExperiment} object with a single matrix of UMI counts.
}
\description{
Obtain the mouse haematopoietic stem cell single-cell RNA-seq data from Grun et al. (2016).
}
\details{
Row metadata contains the symbol and chromosomal location for each gene.
Column metadata contains the extraction protocol used for each sample, as described in GSE76983.

If \code{ensembl=TRUE}, the gene symbols are converted to Ensembl IDs in the row names of the output object.
Rows with missing Ensembl IDs are discarded, and only the first occurrence of duplicated IDs is retained.

If \code{location=TRUE}, the coordinates of the Ensembl gene models are stored in the \code{\link{rowRanges}} of the output.
Note that this is only performed if \code{ensembl=TRUE}.

All data are downloaded from ExperimentHub and cached for local re-use.
Specific resources can be retrieved by searching for \code{scRNAseq/grun-hsc}.
}
\examples{
sce <- GrunHSCData()

}
\references{
Grun D et al. (2016). 
De novo prediction of stem cell identity using single-cell transcriptome data. 
\emph{Cell Stem Cell} 19(2), 266-277.
}
\author{
Aaron Lun
}
