% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{integrate_batches}
\alias{integrate_batches}
\title{Integrate batches}
\usage{
integrate_batches(obj.list)
}
\arguments{
\item{obj.list}{list of the S$ objects to be merged}
}
\value{
Integrated Seurat object
}
\description{
Use the SCT approach of Seurat to integrate batches together
}
