% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaMannoBrainData.R
\name{LaMannoBrainData}
\alias{LaMannoBrainData}
\title{Obtain the La Manno brain data}
\usage{
LaMannoBrainData(which = c("human-es", "human-embryo", "human-ips",
  "mouse-adult", "mouse-embryo"))
}
\arguments{
\item{which}{A string specifying which data set should be obtained.}
}
\value{
A \linkS4class{SingleCellExperiment} object.
}
\description{
Download and cache the La Manno brain single-cell RNA-seq (scRNA-seq) dataset from ExperimentHub,
returning a \linkS4class{SingleCellExperiment} object for further use.
}
\details{
This function provides the brain scRNA-seq data from La Manno et al. (2016)
in the form of a \linkS4class{SingleCellExperiment} object with a single matrix of read counts.

Column metadata is provided in the same form as supplied in the supplementary tables in GSE71585.
This contains information such as the time point and cell type.

The various settings of \code{which} will obtain different data sets.
\itemize{
\item \code{"human-es"}, human embryonic stem cells.
\item \code{"human-embryo"}, human embryo midbrain.
\item \code{"human-ips"}, human induced pluripotent stem cells.
\item \code{"mouse-adult"}, mouse adult dopaminergic neurons.
\item \code{"mouse-embryo"}, mouse embryo midbrain.
}
Unfortunately, each of these datasets uses a different set of features.
If multiple datasets are to be used simultaneously, users will have to decide how to merge them,
e.g., by taking the intersection of common features across all datasets.
}
\examples{
sce <- LaMannoBrainData()

}
\references{
La Manno A et al. (2016).
Molecular diversity of midbrain development in mouse, human, and stem cells. 
\emph{Cell} 167(2), 566-580.
}
\author{
Aaron Lun
}
