% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAvailableVersions.R
\name{listAvailableVersions}
\alias{listAvailableVersions}
\alias{fetchLatestVersion}
\title{List available versions}
\usage{
listAvailableVersions(name, package = "scRNAseq")

fetchLatestVersion(name, package = "scRNAseq")
}
\arguments{
\item{name}{String containing the name of the asset.}

\item{package}{String containing the name of the package.}
}
\value{
For \code{listAvailableVersion}, a character vector containing the names of the available versions of the \code{name} asset.

For \code{fetchLatestVersion}, a string containing the name of the latest version.
}
\description{
List the available and latest versions for an asset (i.e., datasets or collections of datasets).
}
\examples{
listAvailableVersions("zeisel-brain-2015")
fetchLatestVersion("zeisel-brain-2015")

}
\author{
Aaron Lun
}
