% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAvailableDatasets.R
\name{listAvailableDatasets}
\alias{listAvailableDatasets}
\title{List available datasets}
\usage{
listAvailableDatasets(cache = NULL, overwrite = FALSE, latest = TRUE)
}
\arguments{
\item{cache, overwrite}{Arguments to pass to \code{\link[gypsum]{fetchMetadataDatabase}}.}
}
\value{
A \linkS4class{DataFrame} where each row corresponds to a dataset, containing various pieces of metadata.
Some columns may be lists to capture 1:many mappings.
}
\description{
Summary information for all available datasets in the \pkg{scRNAseq} package.
}
\examples{
listAvailableDatasets()

}
\author{
Aaron Lun
}
