% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDatasets.R
\name{listDatasets}
\alias{listDatasets}
\title{List all available datasets}
\usage{
listDatasets()
}
\value{
A \linkS4class{DataFrame} where each row corresponds to a dataset, containing the fields:
\itemize{
\item \code{Reference}, a Markdown-formatted citation to \code{scripts/ref.bib} in the \pkg{scRNAseq} installation directory.
\item \code{Taxonomy}, an identifier for the organism.
\item \code{Part}, the part of the organism being studied.
\item \code{Number}, the total number of cells.
\item \code{Call}, the relevant R call required to generate the dataset.
}
}
\description{
Summary information for all available datasets in the \pkg{scRNAseq} package.
}
\details{
A study may contribute multiple datasets if they cannot be reasonably combined (e.g., different species).
If the authors provided more cells than were used in the analysis, 
the total number of cells prior to author filtering is reported.
}
\examples{
listDatasets()

}
\author{
Aaron Lun
}
