% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_10X}
\alias{load_10X}
\title{Load 10X}
\usage{
load_10X(
  file,
  name = "10X-project",
  percent_mt = 20,
  max_features = 5000,
  min_features = 200
)
}
\arguments{
\item{file}{string (path to file)}

\item{percent_mt}{integer \link{0-100}}

\item{max_features}{integer \link{0-Inf}}

\item{min_features}{integer \link{0-Inf}}
}
\value{
Preprocessed Seurat object
}
\description{
This function takes the path to a 10X output folder and instanciates the Seurat object
}
