% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MacoskoRetinaData.R
\name{MacoskoRetinaData}
\alias{MacoskoRetinaData}
\title{Obtain the Macosko retina data}
\usage{
MacoskoRetinaData()
}
\value{
A \linkS4class{SingleCellExperiment} object.
}
\description{
Download and cache the Macosko retina single-cell RNA-seq (scRNA-seq) dataset from ExperimentHub,
returning a \linkS4class{SingleCellExperiment} object for further use.
}
\details{
This function provides the mouse retina scRNA-seq data from Macosko et al. (2016)
in the form of a \linkS4class{SingleCellExperiment} object with a single matrix of UMI counts.

Column metadata contains the cluster identity as reported in the paper.
Note that some cells will have \code{NA} identities as they are present in the count matrix but not in the metadata file.
These are presumably low-quality cells that were discarded prior to clustering.
}
\examples{
sce <- MacoskoRetinaData()

}
\references{
Macosko E et al. (2016). 
Highly parallel genome-wide expression profiling of individual cells using nanoliter droplets. 
\emph{Cell} 161(5), 1202-1214.
}
\author{
Aaron Lun
}
