% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{merge_samples}
\alias{merge_samples}
\title{Merge samples}
\usage{
merge_samples(files)
}
\arguments{
\item{files}{character vector of the files to merge}
}
\value{
Merged Seurat object
}
\description{
This function allows to instantiate a merged Seurat object of several distinct samples
}
