% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuraroPancreasData.R
\name{MuraroPancreasData}
\alias{MuraroPancreasData}
\title{Obtain the Muraro pancreas data}
\usage{
MuraroPancreasData()
}
\value{
A \linkS4class{SingleCellExperiment} object.
}
\description{
Download and cache the Muraro pancreas single-cell RNA-seq (scRNA-seq) dataset from ExperimentHub,
returning a \linkS4class{SingleCellExperiment} object for further use.
}
\details{
This function provides the human pancreas scRNA-seq data from Muraro et al. (2016)
in the form of a \linkS4class{SingleCellExperiment} object with a single matrix of UMI counts. 

Row data contains fields for the symbol and chromosomal location of each gene.
Spike-ins are specially labelled with the \code{\link{isSpike}} function.

Column metadata is derived from the columns of the count matrix provided in GSE85241,
with additional cell type labels obtained from the authors (indirectly, via the Hemberg group).
Some cells have \code{NA} labels and were presumably removed prior to downstream analyses.
}
\examples{
sce <- MuraroPancreasData()

}
\references{
Muraro MJ et al. (2016). 
A single-cell transcriptome atlas of the human pancreas.
\emph{Cell Syst.} 3(4), 385-394.
}
\author{
Aaron Lun,
using additional metadata obtained by Vladimir Kiselev.
}
