% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{preprocess_10X}
\alias{preprocess_10X}
\title{Preprocess 10X}
\usage{
preprocess_10X(data, percent_mt = 20, max_features = 5000, min_features = 200)
}
\arguments{
\item{data}{data-frame}

\item{percent_mt}{integer \link{0-100}}

\item{max_features}{integer \link{0-Inf}}

\item{min_features}{integer \link{0-Inf}}
}
\value{
Preprocessed Seurat object
}
\description{
This function takes a data-frame (genes x cells),
creates a Seurat object with it and filters the object for default tags such as
Min and max nFeature_RNA and \% of MT
}
