% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recluster}
\alias{recluster}
\title{Recluster}
\usage{
recluster(sub.obj, name = "10X-subcluster", batch = NA)
}
\arguments{
\item{sub.obj}{Seurat object}
}
\value{
Processed reclustering
}
\description{
Takes a subset of a Seurat object and performs reclustering by instanciating a new object
}
