% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{reduce_dim}
\alias{reduce_dim}
\title{Reduce Dimensions}
\usage{
reduce_dim(seurat.obj, ndims = 15, res = 0.1)
}
\arguments{
\item{seurat.obj}{S4 instance}

\item{ndims}{integer \link{3-100}}

\item{res}{double \link{0-3}}
}
\value{
Seurat object with reduction embeddings
}
\description{
This function reduces the dimensions of the Normalized Seurat object
Runs PCA and then UMAP and then performs clustering
}
