% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchDatasets.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{defineTextQuery}
\alias{gsc}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{gypsum}{\code{\link[gypsum:searchMetadata]{defineTextQuery}}, \code{\link[gypsum:searchMetadata]{gsc}}}
}}

