% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RomanovBrainData.R
\name{RomanovBrainData}
\alias{RomanovBrainData}
\title{Obtain the Romanov brain data}
\usage{
RomanovBrainData()
}
\value{
A \linkS4class{SingleCellExperiment} object.
}
\description{
Download and cache the Romanov brain single-cell RNA-seq (scRNA-seq) dataset from ExperimentHub,
returning a \linkS4class{SingleCellExperiment} object for further use.
}
\details{
This function provides the brain scRNA-seq data from Romanov et al. (2017)
in the form of a \linkS4class{SingleCellExperiment} object with a single matrix of read counts.

Column metadata is provided in the same form as supplied in GSE74672.
This contains information such as the reporter gene expressed in each cell, the mouse line, dissection type and so on.

Rows corresponding to spike-in transcripts are labelled with the \code{\link{isSpike}} function.
Note that some of the spike-in rows have \code{NA} observations for some (but not all) cells.
}
\examples{
sce <- RomanovBrainData()

}
\references{
Romanov RA et al. (2017).
Molecular interrogation of hypothalamic organization reveals distinct dopamine neuronal subtypes. 
\emph{Nat. Neurosci.} 20, 176-188.
}
\author{
Aaron Lun,
based on code by Vladimir Kiselev and Tallulah Andrews.
}
