% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveDataset.R
\name{saveDataset}
\alias{saveDataset}
\title{Save a dataset to disk}
\usage{
saveDataset(x, path, metadata)
}
\arguments{
\item{x}{A \linkS4class{SummarizedExperiment} or one of its subclasses.}

\item{path}{String containing the path to a new directory in which to save \code{x}.}

\item{metadata}{Named list containing metadata, usually generated using \code{\link{createMetadata}} or \code{\link{fetchMetadata}}.}
}
\value{
\code{x} and its metadata are saved into \code{path}, and \code{NULL} is invisibly returned.
}
\description{
Save a single-cell dataset to disk, usually in preparation for upload.
}
\examples{
library(SingleCellExperiment)
sce <- SingleCellExperiment(list(counts=matrix(rpois(1000, lambda=1), 100, 10)))
rownames(sce) <- sprintf("GENE_\%i", seq_len(nrow(sce)))
colnames(sce) <- head(LETTERS, 10)

meta <- createMetadata(
    title="My dataset",
    description="This is my dataset",
    taxonomy.id="10090",
    genome="GRCh38",
    sources=list(list(provider="GEO", id="GSE12345")),
    maintainer.name="Shizuka Mogami",
    maintainer.email="mogami.shizuka@765pro.com"
)

tmp <- tempfile()
saveDataset(sce, tmp, meta)
list.files(tmp, recursive=TRUE)
alabaster.base::readObject(tmp)

}
\seealso{
\code{\link{createMetadata}}, to create the metadata.

\code{\link{polishDataset}}, to polish \code{x} before saving it.

\code{\link{uploadDirectory}}, to upload the saved contents.
}
\author{
Aaron Lun
}
