\name{scRNAseq-package}
\alias{scRNAseq-package}
\alias{scRNAseq}
\alias{fluidigm}
\alias{th2}
\alias{allen}

\docType{package}
\title{
\packageTitle{scRNAseq}
}
\description{
\packageDescription{scRNAseq}
}
\details{

This package contains a collection of three publicly available single-cell RNA-seq datasets.

The dataset \code{fluidigm} contains 65 cells from Pollen et al. (2014), each sequenced at high and low coverage.

The dataset \code{th2} contains 96 T helper cells from Mahata et al. (2014).

The dataset \code{allen} contains 379 cells from the mouse visual cortex. This is a subset of the data published in Tasic et al. (2016).

See the package vignette for details on the pre-processing of the data.

}
\author{
\packageAuthor{scRNAseq}

Maintainer: \packageMaintainer{scRNAseq}
}

\references{

Pollen, Nowakowski, Shuga, Wang, Leyrat, Lui, Li, Szpankowski, Fowler, Chen, Ramalingam, Sun, Thu, Norris, Lebofsky, Toppani, Kemp II, Wong, Clerkson, Jones, Wu, Knutsson, Alvarado, Wang, Weaver, May, Jones, Unger, Kriegstein, West. Low-coverage single-cell mRNA sequencing reveals cellular heterogeneity and activated signaling pathways in developing cerebral cortex. Nature Biotechnology, 32, 1053-1058 (2014).

Mahata, Zhang, Kolodziejczyk, Proserpio, Haim-Vilmovsky, Taylor, Hebenstreit, Dingler, Moignard, Gottgens, Arlt, McKenzie, Teichmann. Single-Cell RNA Sequencing Reveals T Helper Cells Synthesizing Steroids De Novo to Contribute to Immune Homeostasis. Cell Reports, 7(4): 1130–1142 (2014).

Tasic, Menon, Nguyen, Kim, Jarsky, Yao, Levi, Gray, Sorensen, Dolbeare, Bertagnolli, Goldy, Shapovalova, Parry, Lee, Smith, Bernard, Madisen, Sunkin, Hawrylycz, Koch, Zeng. Adult mouse cortical cell taxonomy revealed by single cell transcriptomics. Nature Neuroscience, 19, 335–346 (2016).
}
