% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scRNAseq}
\alias{scRNAseq}
\title{scRNAseq}
\format{

}
\source{
\url{https://github.com/matthewhirschey/genes}
}
\usage{
data(scRNAseq)
}
\description{
A data set containing a dummy gene by cell matrix and gene metadata.
}
\keyword{datasets}
