% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScrnaseqArraySeed.R
\name{ScrnaseqArraySeed}
\alias{ScrnaseqArraySeed}
\alias{ScrnaseqArraySeed-class}
\alias{ScrnaseqArray-class}
\alias{ScrnaseqMatrix-class}
\alias{DelayedArray,ScrnaseqArraySeed-method}
\alias{matrixClass,ScrnaseqArray-method}
\alias{coerce,ScrnaseqArray,ScrnaseqMatrix-method}
\alias{coerce,ScrnaseqMatrix,ScrnaseqArray-method}
\alias{path,ScrnaseqArraySeed-method}
\alias{saveObject,ScrnaseqArray-method}
\alias{ScrnaseqArray}
\title{scRNAseq-derived array}
\usage{
ScrnaseqArraySeed(
  name,
  version,
  path,
  cached,
  package = "scRNAseq",
  seed = NULL,
  ...
)

ScrnaseqArray(path, seed = NULL, ...)
}
\arguments{
\item{name}{String containing the name of the asset containing the array.
Alternatively an existing ScrnaseqArraySeed, which is returned without modification.}

\item{version}{String containing the version of the asset containing the array.}

\item{path}{String containing the path to the array inside the asset.}

\item{cached}{String containing the path to the locally cached representation of the array.}

\item{package}{String containing the name of the package containing the asset.}

\item{seed}{Contents of the loaded array.
If \code{NULL}, this is created by calling \code{\link{readObject}} on \code{cached}.}

\item{...}{Further arguments to pass to \code{\link{readObject}} when \code{seed=NULL}.}
}
\value{
For the constructors, an instance of the \linkS4class{ScrnaseqArraySeed} or \linkS4class{ScrnaseqArray}.
}
\description{
A \linkS4class{DelayedArray} class that remembers its provenance, typically from \code{\link{fetchDataset}}.
This allows methods to avoid re-saving the array and uploading a duplicate copy when, e.g., updating an existing dataset.
}
\details{
The ScrnaseqArraySeed is a subclass of the \linkS4class{WrapperArraySeed} and will just forward all operations to the underlying \code{seed}.
Its main purpose is to track the gypsum project-asset-version combination that was originally used to generate \code{seed}, 
which enables optimizations during \code{\link{saveDataset}} and \code{\link{uploadDirectory}}.
}
