% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadDataset.R
\name{uploadDirectory}
\alias{uploadDirectory}
\title{Upload a directory to the gypsum backend}
\usage{
uploadDirectory(
  path,
  name,
  version,
  package = "scRNAseq",
  url = NULL,
  concurrent = 1,
  abort.failed = TRUE
)
}
\arguments{
\item{path}{String containing the path to a directory.
This can either be the same directory generated by \code{\link{saveDataset}}
or it can be a parent directory containing multiple subdirectories generated by \code{\link{saveDataset}}.}

\item{name}{String containing the name of the dataset.
This should not contain \code{/} or start with \code{..}.}

\item{version}{String containing the version of the dataset.
This should not contain \code{/} or start with \code{..}.}

\item{package}{String containing the package name.}

\item{url}{String containing the URL to the gypsum REST API.}

\item{concurrent}{Integer scalar specifying the number of concurrent uploads.}

\item{abort.failed}{Logical scalar indicating whether to abort the upload on any failure.
Setting this to \code{FALSE} can be helpful for diagnosing upload problems.}
}
\value{
On successful upload, \code{NULL} is invisibly returned.
}
\description{
Upload a directory containing one or more datasets to the gypsum backend as a versioned asset.
This requires uploader permissions
}
\examples{
example(saveDataset, verbose=FALSE)
if (interactive()) {
    uploadDirectory(tmp)
}

}
\author{
Aaron Lun
}
