% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrunHSCData.R
\name{GrunHSCData}
\alias{GrunHSCData}
\title{Obtain the Grun HSC data}
\usage{
GrunHSCData()
}
\value{
A \linkS4class{SingleCellExperiment} object.
}
\description{
Download and cache the Grun HSC single-cell RNA-seq (scRNA-seq) dataset from ExperimentHub,
returning a \linkS4class{SingleCellExperiment} object for further use.
}
\details{
This function provides the haematopoietic stem cell scRNA-seq data from Grun et al. (2016)
in the form of a \linkS4class{SingleCellExperiment} object with a single matrix of read counts.

Row metadata contains the symbol and chromosomal location for each gene.
Column metadata contains the extraction protocol used for each sample, as described in GSE76983.
}
\examples{
sce <- GrunHSCData()

}
\references{
Grun D et al. (2016). 
De novo prediction of stem cell identity using single-cell transcriptome data. 
\emph{Cell Stem Cell} 19(2), 266-277.
}
\author{
Aaron Lun
}
