% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrunPancreasData.R
\name{GrunPancreasData}
\alias{GrunPancreasData}
\title{Obtain the Grun pancreas data}
\usage{
GrunPancreasData()
}
\value{
A \linkS4class{SingleCellExperiment} object.
}
\description{
Download and cache the Grun pancreas single-cell RNA-seq (scRNA-seq) dataset from ExperimentHub,
returning a \linkS4class{SingleCellExperiment} object for further use.
}
\details{
This function provides the human pancreas scRNA-seq data from Grun et al. (2016)
in the form of a \linkS4class{SingleCellExperiment} object with a single matrix of UMI counts. 

Row metadata contains fields for the symbol and chromosomal location of each gene,
as derived from the row names.

Column metadata is derived from the column names of the count matrix with the sample annotations in GSE81076.
This includes the donor identity for each cell and the type of sample.

An alternative experiment contains the data for the spike-in transcripts.
}
\examples{
sce <- GrunPancreasData()

}
\references{
Grun D et al. (2016). 
De novo prediction of stem cell identity using single-cell transcriptome data. 
\emph{Cell Stem Cell} 19(2), 266-277.
}
\author{
Aaron Lun,
using additional metadata obtained by Vladimir Kiselev.
}
