% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAvailableDatasets.R
\name{listAvailableDatasets}
\alias{listAvailableDatasets}
\title{List available datasets}
\usage{
listAvailableDatasets(cache = NULL, overwrite = FALSE)
}
\value{
A \linkS4class{DataFrame} where each row corresponds to a dataset, containing various pieces of metadata.
}
\description{
Summary information for all available datasets in the \pkg{scRNAseq} package.
}
\examples{
listAvailableDatasets()

}
\author{
Aaron Lun
}
