% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MacoskoRetinaData.R
\name{MacoskoRetinaData}
\alias{MacoskoRetinaData}
\title{Obtain the Macosko retina data}
\usage{
MacoskoRetinaData(ensembl = FALSE)
}
\arguments{
\item{ensembl}{Logical scalar indicating whether the output row names should contain Ensembl identifiers.}
}
\value{
A \linkS4class{SingleCellExperiment} object with a single matrix of UMI counts.
}
\description{
Obtain the mouse retina single-cell RNA-seq data from Macosko et al. (2016).
}
\details{
Column metadata contains the cluster identity as reported in the paper.
Note that some cells will have \code{NA} identities as they are present in the count matrix but not in the metadata file.
These are presumably low-quality cells that were discarded prior to clustering.

If \code{ensembl=TRUE}, the gene symbols are converted to Ensembl IDs in the row names of the output object.
Rows with missing Ensembl IDs are discarded, and only the first occurrence of duplicated IDs is retained.

All data are downloaded from ExperimentHub and cached for local re-use.
Specific resources can be retrieved by searching for \code{scRNAseq/macosko-retina}.
}
\examples{
sce <- MacoskoRetinaData()

}
\references{
Macosko E et al. (2016). 
Highly parallel genome-wide expression profiling of individual cells using nanoliter droplets. 
\emph{Cell} 161(5), 1202-1214.
}
\author{
Aaron Lun
}
