% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestorowaHSCData.R
\name{NestorowaHSCData}
\alias{NestorowaHSCData}
\title{Obtain the Nestorowa HSC data}
\usage{
NestorowaHSCData(remove.htseq = TRUE)
}
\arguments{
\item{remove.htseq}{Logical scalar indicating whether HT-seq alignment statistics should be removed.}
}
\value{
A \linkS4class{SingleCellExperiment} object.
}
\description{
Download and cache the Nestorowa HSC single-cell RNA-seq (scRNA-seq) dataset from ExperimentHub,
returning a \linkS4class{SingleCellExperiment} object for further use.
}
\details{
This function provides the haematopoietic stem cell scRNA-seq data from Nestorowa et al. (2015)
in the form of a \linkS4class{SingleCellExperiment} object with a single matrix of read counts.

Row corresponding to HT-seq's alignment statistics are present in the count matrix.
By default, these are removed prior to downstream analyses.

Column metadata includes the cell type mapping, as described on the website (see References),
the diffusion map components and the FACS expression levels of selected markers.

ERCC spike-ins are stored as alternative experiments.
}
\examples{
sce <- NestorowaHSCData()

}
\references{
Nestorowa S et al. (2016). 
A single-cell resolution map of mouse hematopoietic stem and progenitor cell differentiation
\emph{Blood} 128, e20-e31.

Gene and protein expression in adult haematopoiesis: Data.
\url{http://blood.stemcells.cam.ac.uk/single_cell_atlas.html#data}.
}
\author{
Aaron Lun
}
