% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegerstolpePancreasData.R
\name{SegerstolpePancreasData}
\alias{SegerstolpePancreasData}
\title{Obtain the Segerstolpe pancreas data}
\usage{
SegerstolpePancreasData()
}
\value{
A \linkS4class{SingleCellExperiment} object.
}
\description{
Download and cache the Segerstolpe pancreas single-cell RNA-seq (scRNA-seq) dataset from ExperimentHub,
returning a \linkS4class{SingleCellExperiment} object for further use.
}
\details{
This function provides the pancreas scRNA-seq data from Segerstolpe et al. (2016)
in the form of a \linkS4class{SingleCellExperiment} object with a single matrix of read counts.

Row data contains fields for the gene symbol and RefSeq transcript IDs corresponding to each gene.
Spike-ins are specially labelled with the \code{\link{isSpike}} function.
The rows of the output object are named with the symbol, but note that these are not unique.

Column metadata were extracted from the \code{Characteristcs} fields of the SDRF file for ArrayExpress E-MTAB-5061.
This contains information such as the cell type labels and patient status.
}
\examples{
sce <- SegerstolpePancreasData()

}
\references{
Segerstolpe A et al. (2016). 
Single-cell transcriptome profiling of human pancreatic islets in health and type 2 diabetes. 
\emph{Cell Metab.} 24(4), 593-607.
}
\author{
Aaron Lun
}
