% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadDirectory.R
\name{uploadDirectory}
\alias{uploadDirectory}
\title{Upload a directory to the gypsum backend}
\usage{
uploadDirectory(
  dir,
  name,
  version,
  package = "scRNAseq",
  url = NULL,
  probation = FALSE,
  concurrent = 1,
  abort.failed = TRUE
)
}
\arguments{
\item{dir}{String containing the path to a directory.
This can either be the same directory generated by \code{\link{saveDataset}}
or it can be a parent directory containing multiple subdirectories generated by \code{\link{saveDataset}}.}

\item{name}{String containing the name of the asset. 
This should not contain \code{/} or start with \code{..}.}

\item{version}{String containing the version of the asset. 
This should not contain \code{/} or start with \code{..}.}

\item{package}{String containing the package name.}

\item{url}{String containing the URL to the gypsum REST API.}

\item{probation}{Logical scalar indicating whether this should be a probational upload.}

\item{concurrent}{Integer scalar specifying the number of concurrent uploads.}

\item{abort.failed}{Logical scalar indicating whether to abort the upload on any failure.
Setting this to \code{FALSE} can be helpful for diagnosing upload problems.}
}
\value{
On successful upload, \code{NULL} is invisibly returned.
}
\description{
Upload a directory containing one or more datasets to the gypsum backend as a versioned asset.
This requires uploader permissions to the \pkg{scRNAseq} project.
}
\details{
\code{dir} may contain multiple related datasets in subdirectories, 
generated by \code{\link{saveDataset}} calls on different \linkS4class{SummarizedExperiment} objects.
This is occasionally useful for studies with multiple outputs, e.g., for different species or modalities.

The \code{probation} flag allows contributors to test their uploads 
and for the \pkg{scRNAseq} maintainers to review the content before approval.
Note that uploads from untrusted users are automatically probational so setting \code{probation=FALSE} is not explicitly required in such cases.
}
\examples{
library(SingleCellExperiment)
sce <- SingleCellExperiment(list(counts=matrix(rpois(1000, lambda=1), 100, 10)))
rownames(sce) <- sprintf("GENE_\%i", seq_len(nrow(sce)))
colnames(sce) <- head(LETTERS, 10)

meta <- createMetadata(
    title="My dataset",
    description="This is my dataset",
    taxonomy.id="10090",
    genome="GRCh38",
    sources=list(list(provider="GEO", id="GSE12345")),
    maintainer.name="Shizuka Mogami",
    maintainer.email="mogami.shizuka@765pro.com"
)

cache <- tempfile()
saveDataset(sce, cache, meta)

version <- as.character(Sys.Date())
if (interactive()) {
    # Uploading a probational version for test purposes.
    uploadDirectory(cache, "test", version, probation=TRUE)

    # Cleaning up after ourselves.
    gypsum::rejectProbation("scRNAseq", "test", version)
}

}
\seealso{
\code{\link{saveDataset}}, to save data to a directory.

\code{\link{fetchDataset}}, to download an existing dataset into the current sesssion.
}
\author{
Aaron Lun
}
