% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsoskinBrainData.R
\name{UsoskinBrainData}
\alias{UsoskinBrainData}
\title{Obtain the Usoskin brain data}
\usage{
UsoskinBrainData()
}
\value{
A \linkS4class{SingleCellExperiment} object.
}
\description{
Download and cache the Usoskin brain single-cell RNA-seq (scRNA-seq) dataset from ExperimentHub,
returning a \linkS4class{SingleCellExperiment} object for further use.
}
\details{
This function provides the brain scRNA-seq data from Usoskin et al. (2015)
in the form of a \linkS4class{SingleCellExperiment} object with a single matrix of RPMs.

Column metadata is provided in the same form as supplied in External Table 2 of \url{http://linnarssonlab.org/drg/}.
This contains information such as the library of origin and the cell type.

The count matrix contains information for repeats, marked with \code{r_} prefixes in the row names;
as well as mitochondrial transcripts, marked with \code{mt-} prefixes.
}
\examples{
sce <- UsoskinBrainData()

}
\references{
Usoskin A et al. (2015).
Unbiased classification of sensory neuron types by large-scale single-cell RNA sequencing. 
\emph{Nat. Neurosci.} 18(1), 145-53.
}
\author{
Aaron Lun
}
