% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XinPancreasData.R
\name{XinPancreasData}
\alias{XinPancreasData}
\title{Obtain the Xin pancreas data}
\usage{
XinPancreasData()
}
\value{
A \linkS4class{SingleCellExperiment} object.
}
\description{
Download and cache the Xin pancreas single-cell RNA-seq (scRNA-seq) dataset from ExperimentHub,
returning a \linkS4class{SingleCellExperiment} object for further use.
}
\details{
This function provides the pancreas scRNA-seq data from Xin et al. (2016)
in the form of a \linkS4class{SingleCellExperiment} object with a single matrix of RPKMs.

Row data contains fields for the Entrez ID and symbol for each gene.
Column metadata was obtained from the authors (indirectly, via the Hemberg group) 
and contains information such as the cell type labels and donor status.
}
\examples{
sce <- XinPancreasData()

}
\references{
Xin A et al. (2016).
RNA sequencing of single human islet cells reveals type 2 diabetes genes.
\emph{Cell Metab.} 24(4), 608-615.
}
\author{
Aaron Lun,
using additional metadata obtained by Vladimir Kiselev.
}
