% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZeiselBrainData.R
\name{ZeiselBrainData}
\alias{ZeiselBrainData}
\title{Obtain the Zeisel brain data}
\usage{
ZeiselBrainData()
}
\value{
A \linkS4class{SingleCellExperiment} object.
}
\description{
Download and cache the Zeisel brain single-cell RNA-seq (scRNA-seq) dataset from ExperimentHub,
returning a \linkS4class{SingleCellExperiment} object for further use.
}
\details{
This function provides the brain scRNA-seq data from Zeisel et al. (2015)
in the form of a \linkS4class{SingleCellExperiment} object with a single matrix of UMI counts.

Row data contains a single \code{"featureType"} field describing the type of each feature
(endogenous genes, mitochondrial genes, spike-in transcripts and repeats).
Spike-ins are also specially labelled with the \code{\link{isSpike}} function.

Column metadata is provided in the same form as supplied in \url{http://linnarssonlab.org/cortex/}.
This contains information such as the cell diameter and the published cell type annotations.
}
\examples{
sce <- ZeiselBrainData()

}
\references{
Zeisel A et al. (2015). 
Brain structure. Cell types in the mouse cortex and hippocampus revealed by single-cell RNA-seq. 
\emph{Science} 347(6226), 1138-42.
}
\author{
Aaron Lun
}
